fix: Move more classes around
