+++ Assets/Scripts/RidersGame/StuntSystem/StuntDatabase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dbb94b31ccd651241867fed850f52359
